FingerColliders test added
